Update and mange a peer's networking rule on the agent
- Move related functions from the controller to the agent
- It can simplify networking rule update process, and also
- reduce the number of I/Os and amount of data to etcd.
